ajustes no alterar

+++ src/br/com/devti/gestaotransportadora/service/UsuarioService.java
@@ -2,10 +2,8 @@ package br.com.devti.gestaotransportadora.service;
 
 import java.util.List;
 
-import br.com.devti.gestaotransportadora.entity.ClienteEntity;
 import br.com.devti.gestaotransportadora.entity.UsuarioEntity;
 import br.com.devti.gestaotransportadora.util.exception.NegocioException;
-import br.com.devtigestaotransportadora.bo.ClienteBO;
 import br.com.devtigestaotransportadora.bo.UsuarioBO;
 
 public class UsuarioService {
@@ -24,7 +22,7 @@ public class UsuarioService {
         return new UsuarioBO().listarUsuarios();
     }
 
-    public void excluirUsuario(Long id) throws NegocioException {
+    public void excluirUsuario(Integer id) throws NegocioException {
         new UsuarioBO().excluirUsuario(id);
     }
 
@@ -32,7 +30,7 @@ public class UsuarioService {
         return new UsuarioBO().alterarUsuario(usuario);
     }
 
-    public UsuarioEntity buscarUsuarioPorId(Long id) throws NegocioException {
+    public UsuarioEntity buscarUsuarioPorId(Integer id) throws NegocioException {
         return new UsuarioBO().buscarUsuarioPorId(id);
     }
